requirements.txt
bcrypt==3.1.7
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
packaging==20.3
Pillow==7.1.1
SQLAlchemy==1.3.15
virtualenv==20.0.18
WTForms==2.2.1
